Provide a combined upgrade script
This commit provides a new "upgrade_NEXT.sql" file that will become the
combined upgrade script for the next version of this module. Before the
next version can be released, this file needs to be renamed to include
the version number (e.g. upgrade_1_2.sql) and the upgrade process
needs to be tested to make sure it processes the new file correctly.

Signed-off-by: Remington Steed <rjs7@calvin.edu>